Update vehicle buy, sell, chopchop, unimpound
