+++ Altis_Life.Altis/Config_Master.hpp
@@ -74,7 +74,11 @@ class Life_Settings {
 
     /* Car-shop Settings */
     vehicleShop_rentalOnly[] = { "B_MRAP_01_hmg_F", "B_G_Offroad_01_armed_F", "B_Boat_Armed_01_minigun_F" };
-
+	permanentbuymultiplikator = 1.5;
+	garagesellmultiplikator = 0.5;
+	storagefeemultiplikator = 0.2;
+	chopShopmultiplikator = 0.25;
+	
     /* Job-related stuff */
     delivery_points[] = { "dp_1", "dp_2", "dp_3", "dp_4", "dp_5", "dp_6", "dp_7", "dp_8", "dp_9", "dp_10", "dp_11", "dp_12", "dp_13", "dp_14", "dp_15", "dp_15", "dp_16", "dp_17", "dp_18", "dp_19", "dp_20", "dp_21", "dp_22", "dp_23", "dp_24", "dp_25" };
 
@@ -130,6 +134,8 @@ class Life_Settings {
 
 #include "Config_Clothing.hpp"
 #include "Config_Licenses.hpp"
-#include "Config_Vehicles.hpp"
+#include "Config_VehicleConfig.hpp"
+#include "Config_VehicleShop.hpp"
 #include "Config_vItems.hpp"
 #include "Config_Weapons.hpp"
+
