+++ Altis_Life.Altis/Config_VehicleConfig.hpp
@@ -1,0 +1,522 @@
+class LifeCfgVehicles {
+    /*
+    *    Vehicle Configs (Contains textures and other stuff)
+    *	 1: ARRAY (Rental Price)
+	*	 Ex: { 200, 200 , 200 , 200 } //civilian, west, independent, east
+    *    2: ARRAY (license required)
+    *    Ex: { "driver", "" , "" , "" } //civilian, west, independent, east
+    *
+    *    Textures config follows { Texture Name, side, {texture(s)path}}
+    *    Texture(s)path follows this format:
+    *    INDEX 0: Texture Layer 0
+    *    INDEX 1: Texture Layer 1
+    *    INDEX 2: Texture Layer 2
+    *    etc etc etc
+    */
+
+    class Default {
+        vItemSpace = -1;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { -1, -1, -1, -1 };
+        textures[] = {};
+    };
+
+    class I_Truck_02_medical_F {
+        vItemSpace = 150;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { -1, -1, 25000, -1 };
+        textures[] = {};
+    };
+
+    class O_Truck_03_medical_F {
+        vItemSpace = 200;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { -1, -1, 45000, -1 };
+        textures[] = {};
+    };
+
+    class B_Truck_01_medical_F {
+        vItemSpace = 250;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { -1, -1, 60000, -1 };
+        textures[] = {};
+    };
+
+    class C_Rubberboat {
+        vItemSpace = 45;
+		licenses[] = { { "boat"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 5000, -1, -1, -1 };
+        textures[] = { };
+    };
+
+    class B_Heli_Transport_01_F {
+        vItemSpace = 200;
+		licenses[] = { { ""}, { "cAir" }, { "" }, { "" }};
+		rentalprice[] = { -1, 200000, -1, -1 };
+        textures[] = {};
+    };
+
+    class B_MRAP_01_hmg_F {
+        vItemSpace = 100;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { -1, 750000, -1, -1 };
+        textures[] = {
+            { "Black", "cop", {
+                "#(argb,8,8,3)color(0.05,0.05,0.05,1)",
+                "#(argb,8,8,3)color(0.05,0.05,0.05,1)",
+                "#(argb,8,8,3)color(0.05,0.05,0.05,1)"
+            } }
+        };
+    };
+
+    class B_Boat_Armed_01_minigun_F {
+        vItemSpace = 175;
+		licenses[] = { { ""}, { "cg" }, { "" }, { "" }};
+		rentalprice[] = { -1, 75000, -1, -1 };
+        textures[] = { };
+    };
+
+    class B_Boat_Transport_01_F {
+        vItemSpace = 45;
+		licenses[] = { { ""}, { "cg" }, { "" }, { "" }};
+		rentalprice[] = { -1, 3000, -1, -1 };
+        textures[] = { };
+    };
+
+    class O_Truck_03_transport_F {
+        vItemSpace = 285;
+		licenses[] = { { "trucking"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 200000, -1, -1, -1 };
+        textures[] = { };
+    };
+
+    class O_Truck_03_device_F {
+        vItemSpace = 350;
+		licenses[] = { { "trucking"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 450000, -1, -1, -1 };
+        textures[] = { };
+    };
+
+    class Land_CargoBox_V1_F {
+        vItemSpace = 5000;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { -1, -1, -1, -1 };
+        textures[] = {};
+    };
+
+    class Box_IND_Grenades_F {
+        vItemSpace = 350;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { -1, -1, -1, -1 };
+        textures[] = {};
+    };
+
+    class B_supplyCrate_F {
+        vItemSpace = 700;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { -1, -1, -1, -1 };
+        textures[] = {};
+    };
+
+    class B_G_Offroad_01_F {
+        vItemSpace = 65;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 12500, -1, -1, -1 };
+        textures[] = { };
+    };
+
+    class B_G_Offroad_01_armed_F {
+        vItemSpace = 65;
+		licenses[] = { { "rebel"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 750000, -1, -1, -1 };
+        textures[] = { };
+    };
+
+    class C_Boat_Civil_01_F {
+        vItemSpace = 85;
+		licenses[] = { { "boat"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 22000, -1, -1, -1 };
+        textures[] = { };
+    };
+
+    class C_Boat_Civil_01_police_F {
+        vItemSpace = 85;
+		licenses[] = { { ""}, { "cg" }, { "" }, { "" }};
+		rentalprice[] = { -1, 20000, -1, -1 };
+        textures[] = { };
+    };
+
+    class B_Truck_01_box_F {
+        vItemSpace = 450;
+		licenses[] = { { "trucking"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 350000, -1, -1, -1 };
+        textures[] = { };
+    };
+
+    class B_Truck_01_transport_F {
+        vItemSpace = 325;
+		licenses[] = { { "trucking"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 275000, -1, -1, -1 };
+        textures[] = { };
+    };
+
+    class O_MRAP_02_F {
+        vItemSpace = 60;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 150000, -1, -1, -1 };
+        textures[] = { };
+    };
+
+    class C_Offroad_01_F {
+        vItemSpace = 65;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 12500, 5000, 10000, -1 };
+        textures[] = {
+            { "Red", "civ", {
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_co.paa",
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_co.paa"
+            } },
+            { "Yellow", "civ", {
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE01_CO.paa",
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE01_CO.paa"
+            } },
+            { "White", "civ", {
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE02_CO.paa",
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE02_CO.paa"
+            } },
+            { "Blue", "civ", {
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE03_CO.paa",
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE03_CO.paa"
+            } },
+            { "Dark Red", "civ", {
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE04_CO.paa",
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE04_CO.paa"
+            } },
+            { "Blue / White", "civ", {
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE05_CO.paa",
+                "\A3\soft_F\Offroad_01\Data\offroad_01_ext_BASE05_CO.paa"
+            } },
+            { "Taxi", "civ", {
+                "#(argb,8,8,3)color(0.6,0.3,0.01,1)"
+            } },
+            { "Police", "cop", {
+                "#(ai,64,64,1)Fresnel(1.3,7)"
+            } }
+        };
+    };
+
+    class C_Kart_01_Blu_F {
+        vItemSpace = 20;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 15000, -1, -1, -1 };
+        textures[] = {};
+    };
+
+    class C_Kart_01_Fuel_F {
+        vItemSpace = 20;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 15000, -1, -1, -1 };
+        textures[] = {};
+    };
+
+    class C_Kart_01_Red_F {
+        vItemSpace = 20;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 15000, -1, -1, -1 };
+        textures[] = {};
+    };
+
+    class C_Kart_01_Vrana_F {
+        vItemSpace = 20;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 15000, -1, -1, -1 };
+        textures[] = {};
+    };
+
+    class C_Hatchback_01_sport_F {
+        vItemSpace = 45;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 40000, 30000, -1, -1 };
+        textures[] = {
+            { "Red", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_sport01_co.paa"
+            } },
+            { "Dark Blue", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_sport02_co.paa"
+            } },
+            { "Orange", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_sport03_co.paa"
+            } },
+            { "Black / White", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_sport04_co.paa"
+            } },
+            { "Beige", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_sport05_co.paa"
+            } },
+            { "Green", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_sport06_co.paa"
+            } },
+            { "Police", "cop", {
+                "#(ai,64,64,1)Fresnel(1.3,7)"
+            } }
+        };
+    };
+
+    class B_Quadbike_01_F {
+        vItemSpace = 25;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 2500, -1, -1, -1 };
+        textures[] = {
+            { "Brown", "cop", {
+                "\A3\Soft_F\Quadbike_01\Data\Quadbike_01_co.paa"
+            } },
+            { "Digi Desert", "reb", {
+                "\A3\Soft_F\Quadbike_01\Data\quadbike_01_opfor_co.paa"
+            } },
+            { "Black", "civ", {
+                "\A3\Soft_F_beta\Quadbike_01\Data\quadbike_01_civ_black_co.paa"
+            } },
+            { "Blue", "civ", {
+                "\A3\Soft_F_beta\Quadbike_01\Data\quadbike_01_civ_blue_co.paa"
+            } },
+            { "Red", "civ", {
+                "\A3\Soft_F_beta\Quadbike_01\Data\quadbike_01_civ_red_co.paa"
+            } },
+            { "White", "civ", {
+                "\A3\Soft_F_beta\Quadbike_01\Data\quadbike_01_civ_white_co.paa"
+            } },
+            { "Digi Green", "civ", {
+                "\A3\Soft_F_beta\Quadbike_01\Data\quadbike_01_indp_co.paa"
+            } },
+            { "Hunter Camo", "civ", {
+                "\a3\soft_f_gamma\Quadbike_01\data\quadbike_01_indp_hunter_co.paa"
+            } },
+            { "Rebel Camo", "reb", {
+                "\a3\soft_f_gamma\Quadbike_01\data\quadbike_01_indp_hunter_co.paa"
+            } }
+        };
+    };
+
+    class I_Truck_02_covered_F {
+        vItemSpace = 250;
+		licenses[] = { { "trucking"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 100000, -1, -1, -1 };
+        textures[] = {
+            { "Orange", "civ", {
+                "\A3\Soft_F_Beta\Truck_02\data\truck_02_kab_co.paa",
+                "\a3\soft_f_beta\Truck_02\data\truck_02_kuz_co.paa"
+            } },
+            { "Black", "cop", {
+                "#(argb,8,8,3)color(0.05,0.05,0.05,1)"
+            } }
+        };
+    };
+
+    class I_Truck_02_transport_F {
+        vItemSpace = 200;
+		licenses[] = { { "trucking"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 75000, -1, -1, -1 };
+        textures[] = {
+            { "Orange", "civ", {
+                "\A3\Soft_F_Beta\Truck_02\data\truck_02_kab_co.paa",
+                "\a3\soft_f_beta\Truck_02\data\truck_02_kuz_co.paa"
+            } },
+            { "Black", "cop", {
+                "#(argb,8,8,3)color(0.05,0.05,0.05,1)"
+            } }
+        };
+    };
+
+    class O_Truck_03_covered_F {
+        vItemSpace = 300;
+		licenses[] = { { "trucking"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 250000, -1, -1, -1 };
+        textures[] = {};
+    };
+
+    class C_Hatchback_01_F {
+        vItemSpace = 40;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 9500, -1, -1, -1 };
+        textures[] = {
+            { "Beige", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_base01_co.paa"
+            } },
+            { "Green", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_base02_co.paa"
+            } },
+            { "Blue", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_base03_co.paa"
+            } },
+            { "Dark Blue", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_base04_co.paa"
+            } },
+            { "Yellow", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_base06_co.paa"
+            } },
+            { "White", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_base07_co.paa"
+            } },
+            { "Grey", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_base08_co.paa"
+            } },
+            { "Black", "civ", {
+                "\a3\soft_f_gamma\Hatchback_01\data\hatchback_01_ext_base09_co.paa"
+            } }
+        };
+    };
+
+    class C_SUV_01_F {
+        vItemSpace = 50;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 30000, 20000, -1, -1 };
+        textures[] = {
+            { "Dark Red", "civ", {
+                "\a3\soft_f_gamma\SUV_01\Data\suv_01_ext_co.paa"
+            } },
+            { "Silver", "civ", {
+                "\a3\soft_f_gamma\SUV_01\Data\suv_01_ext_03_co.paa"
+            } },
+            { "Orange", "civ", {
+                "\a3\soft_f_gamma\SUV_01\Data\suv_01_ext_04_co.paa"
+            } },
+            { "Police", "cop", {
+                "\a3\soft_f_gamma\SUV_01\Data\suv_01_ext_02_co.paa"
+            } },
+        };
+    };
+
+    class C_Van_01_transport_F {
+        vItemSpace = 100;
+		licenses[] = { { "driver"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 45000, -1, -1, -1 };
+        textures[] = {
+            { "White", "civ", {
+                "\a3\soft_f_gamma\Van_01\Data\van_01_ext_co.paa"
+            } },
+            { "Red", "civ", {
+                "\a3\soft_f_gamma\Van_01\Data\van_01_ext_red_co.paa"
+            } }
+        };
+    };
+
+    class C_Van_01_box_F {
+        vItemSpace = 150;
+		licenses[] = { { "trucking"}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 60000, -1, -1, -1 };
+        textures[] = {
+            { "White", "civ", {
+                "\a3\soft_f_gamma\Van_01\Data\van_01_ext_co.paa"
+            } },
+            { "Red", "civ", {
+                "\a3\soft_f_gamma\Van_01\Data\van_01_ext_red_co.paa"
+            } }
+        };
+    };
+
+    class B_MRAP_01_F {
+        vItemSpace = 65;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { -1, 30000, -1, -1 };
+        textures[] = {
+            { "Black", "cop", {
+                "#(argb,8,8,3)color(0.05,0.05,0.05,1)",
+                "#(argb,8,8,3)color(0.05,0.05,0.05,1)"
+            } }
+        };
+    };
+	
+	 class B_Heli_Light_01_stripped_F {
+        vItemSpace = 90;
+		licenses[] = { { ""}, { "" }, { "" }, { "" }};
+		rentalprice[] = { 325000, -1, -1, -1 };
+        textures[] = {
+            { "Rebel Digital", "reb", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_digital_co.paa"
+            } }
+        };
+    };
+	
+    class B_Heli_Light_01_F {
+        vItemSpace = 90;
+		licenses[] = { { "pilot"}, { "cAir" }, { "mAir" }, { "" }};
+		rentalprice[] = { 275000, 75000, 50000, -1 };
+        textures[] = {
+            { "Police", "cop", {
+                "\a3\air_f\Heli_Light_01\Data\heli_light_01_ext_ion_co.paa"
+            } },
+            { "Sheriff", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_sheriff_co.paa"
+            } },
+            { "Civ Blue", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\heli_light_01_ext_blue_co.paa"
+            } },
+            { "Civ Red", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\heli_light_01_ext_co.paa"
+            } },
+            { "Blueline", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_blueline_co.paa"
+            } },
+            { "Elliptical", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_elliptical_co.paa"
+            } },
+            { "Furious", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_furious_co.paa"
+            } },
+            { "Jeans Blue", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_jeans_co.paa"
+            } },
+            { "Speedy Redline", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_speedy_co.paa"
+            } },
+            { "Sunset", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_sunset_co.paa"
+            } },
+            { "Vrana", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_vrana_co.paa"
+            } },
+            { "Waves Blue", "civ", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_wave_co.paa"
+            } },
+            { "Rebel Digital", "reb", {
+                "\a3\air_f\Heli_Light_01\Data\Skins\heli_light_01_ext_digital_co.paa"
+            } },
+            { "Digi Green", "reb", {
+                "\a3\air_f\Heli_Light_01\Data\heli_light_01_ext_indp_co.paa"
+            } },
+            { "EMS White", "med", {
+                "#(argb,8,8,3)color(1,1,1,0.8)"
+            } }
+        };
+    };
+
+    class O_Heli_Light_02_unarmed_F {
+        vItemSpace = 210;
+		licenses[] = { { "pilot" }, { "" }, { "mAir" }, { "" }};
+		rentalprice[] = { 750000, -1, 75000, -1 };
+        textures[] = {
+            { "Black", "cop", {
+                "\a3\air_f\Heli_Light_02\Data\heli_light_02_ext_co.paa"
+            } },
+            { "White / Blue", "civ", {
+                "\a3\air_f\Heli_Light_02\Data\heli_light_02_ext_civilian_co.paa"
+            } },
+            { "Digi Green", "civ", {
+                "\a3\air_f\Heli_Light_02\Data\heli_light_02_ext_indp_co.paa"
+            } },
+            { "Desert Digi", "reb", {
+                "\a3\air_f\Heli_Light_02\Data\heli_light_02_ext_opfor_co.paa"
+            } },
+            { "EMS White", "med", {
+                "#(argb,8,8,3)color(1,1,1,0.8)"
+            } }
+        };
+    };
+
+    class B_SDV_01_F {
+        vItemSpace = 50;
+		licenses[] = { { "boat"}, { "cg" }, { "" }, { "" }};
+		rentalprice[] = { 150000, 100000, -1, -1 };
+        textures[] = {};
+    };
+};
