+++ Altis_Life.Altis/Config_VehicleShop.hpp
@@ -1,0 +1,125 @@
+class CarShops {
+    /*
+    *    ARRAY FORMAT:
+    *        0: STRING (Classname)
+    *        1: ARRAY (This is for limiting items to certain things)
+    *            0: Variable to read from
+    *            1: Variable Value Type (SCALAR / BOOL /EQUAL)
+    *            2: What to compare to (-1 = Check Disabled)
+    *
+    *   BLUFOR Vehicle classnames can be found here: https://community.bistudio.com/wiki/Arma_3_CfgVehicles_WEST
+    *   OPFOR Vehicle classnames can be found here: https://community.bistudio.com/wiki/Arma_3_CfgVehicles_EAST
+    *   Independent Vehicle classnames can be found here: https://community.bistudio.com/wiki/Arma_3_CfgVehicles_GUER
+    *   Civilian Vehicle classnames can be found here: https://community.bistudio.com/wiki/Arma_3_CfgVehicles_CIV
+    */
+    class civ_car {
+        side = "civ";
+        vehicles[] = {
+            { "B_Quadbike_01_F", { "", "", -1 } },
+            { "C_Hatchback_01_F", { "", "", -1 } },
+            { "C_Offroad_01_F", { "", "", -1 } },
+            { "C_SUV_01_F", { "", "", -1 } },
+            { "C_Hatchback_01_sport_F", { "", "", -1 } },
+            { "C_Van_01_transport_F", { "", "", -1 } }
+        };
+    };
+
+    class kart_shop {
+        side = "civ";
+        vehicles[] = {
+            { "C_Kart_01_Blu_F", { "", "", -1 } },
+            { "C_Kart_01_Fuel_F", { "", "", -1 } },
+            { "C_Kart_01_Red_F", { "", "", -1 } },
+            { "C_Kart_01_Vrana_F", { "", "", -1 } }
+        };
+    };
+	
+    class civ_truck {
+        side = "civ";
+        vehicles[] = {
+            { "C_Van_01_box_F", { "", "", -1 } },
+            { "I_Truck_02_transport_F", { "", "", -1 } },
+            { "I_Truck_02_covered_F", { "", "", -1 } },
+            { "B_Truck_01_transport_F", { "", "", -1 } },
+            { "O_Truck_03_transport_F", { "", "", -1 } },
+            { "O_Truck_03_covered_F", { "", "", -1 } },
+            { "B_Truck_01_box_F", { "", "", -1 } },
+            { "O_Truck_03_device_F", { "", "", -1 } }
+        };
+    };
+	
+	class civ_air {
+        side = "civ";
+        vehicles[] = {
+            { "B_Heli_Light_01_F", { "", "", -1 } },
+            { "O_Heli_Light_02_unarmed_F", { "", "", -1 } }
+        };
+    };
+	
+	 class civ_ship {
+        side = "civ";
+        vehicles[] = {
+            { "C_Rubberboat", { "", "", -1 } },
+            { "C_Boat_Civil_01_F", { "", "", -1 } },
+            { "B_SDV_01_F", { "", "", -1 } }
+        };
+    };
+
+    class reb_car {
+        side = "civ";
+        vehicles[] = {
+            { "B_Quadbike_01_F", { "", "", -1 } },
+            { "B_G_Offroad_01_F", { "", "", -1 } },
+            { "O_MRAP_02_F", { "", "", -1 } },
+            { "B_Heli_Light_01_stripped_F", { "", "", -1 } },
+            { "B_G_Offroad_01_armed_F", { "", "", -1 } }
+        };
+    };
+	
+	class med_shop {
+        side = "med";
+        vehicles[] = {
+            { "C_Offroad_01_F", { "", "", -1 } },
+            { "I_Truck_02_medical_F", { "", "", -1 } },
+            { "O_Truck_03_medical_F", { "", "", -1 } },
+            { "B_Truck_01_medical_F", { "", "", -1 } }
+        };
+    };
+
+    class med_air_hs {
+        side = "med";
+        vehicles[] = {
+            { "B_Heli_Light_01_F", { "", "", -1 } },
+            { "O_Heli_Light_02_unarmed_F", { "", "", -1 } }
+        };
+    };
+
+	class cop_car {
+        side = "cop";
+        vehicles[] = {
+            { "C_Offroad_01_F", { "", "", -1 } },
+            { "C_SUV_01_F", { "", "", -1 } },
+            { "C_Hatchback_01_sport_F", { "life_coplevel", "SCALAR", 1 } },
+            { "B_MRAP_01_F", { "life_coplevel", "SCALAR", 2 } },
+            { "B_MRAP_01_hmg_F", { "life_coplevel", "SCALAR", 3 } }
+        };
+    };
+
+    class cop_air {
+        side = "cop";
+        vehicles[] = {
+            { "B_Heli_Light_01_F", { "", "", -1 } },
+            { "B_Heli_Transport_01_F", { "life_coplevel", "SCALAR", 3 } }
+        };
+    };
+
+    class cop_ship {
+        side = "cop";
+        vehicles[] = {
+            { "B_Boat_Transport_01_F", { "", "", -1 } },
+            { "C_Boat_Civil_01_police_F", { "", "", -1 } },
+            { "B_Boat_Armed_01_minigun_F", { "life_coplevel", "SCALAR", 3 } },
+            { "B_SDV_01_F", { "", "", -1 } }
+        };
+    };
+};
