+++ Altis_Life.Altis/core/shops/fn_vehicleShopLBChange.sqf
@@ -18,12 +18,22 @@ _className = _control lbData _index;
 _classNameLife = _className;
 _vIndex = _control lbValue _index;
 
-_vehicleList = M_CONFIG(getArray,"CarShops",SEL(life_veh_shop,0),"vehicles");
-_basePrice = SEL(SEL(_vehicleList,_vIndex),1);
+_basePrice = switch(playerSide) do {
+	case civilian: {SEL(M_CONFIG(getArray,CONFIG_LIFE_VEHICLES,_classNameLife,"rentalprice"),0)};
+	case west: {SEL(M_CONFIG(getArray,CONFIG_LIFE_VEHICLES,_classNameLife,"rentalprice"),1)};
+	case independent: {SEL(M_CONFIG(getArray,CONFIG_LIFE_VEHICLES,_classNameLife,"rentalprice"),2)};
+	case east: {SEL(M_CONFIG(getArray,CONFIG_LIFE_VEHICLES,_classNameLife,"rentalprice"),3)};
+};
 
 _vehicleInfo = [_className] call life_fnc_fetchVehInfo;
 _trunkSpace = [_className] call life_fnc_vehicleWeightCfg;
+_maxspeed = _vehicleInfo select 8;
+_horsepower = _vehicleInfo select 11;
+_passengerseats = _vehicleInfo select 10;
+_fuel = _vehicleInfo select 12;
+_armor = _vehicleInfo select 9;
 [_className] call life_fnc_vehicleShop3DPreview;
+_vehiclepermanentbuy = LIFE_SETTINGS(getNumber,"permanentbuymultiplikator");
 
 ctrlShow [2330,true];
 (CONTROL(2300,2303)) ctrlSetStructuredText parseText format[
@@ -33,16 +43,16 @@ ctrlShow [2330,true];
 (localize "STR_Shop_Veh_UI_HPower")+ " %4<br/>" +
 (localize "STR_Shop_Veh_UI_PSeats")+ " %5<br/>" +
 (localize "STR_Shop_Veh_UI_Trunk")+ " %6<br/>" +
-(localize "STR_Shop_Veh_UI_Fuel")+ " %7l<br/>" +
+(localize "STR_Shop_Veh_UI_Fuel")+ " %7<br/>" +
 (localize "STR_Shop_Veh_UI_Armor")+ " %8",
 [_basePrice] call life_fnc_numberText,
-[round(_basePrice * 1.5)] call life_fnc_numberText,
-_vehicleInfo select 8,
-_vehicleInfo select 11,
-_vehicleInfo select 10,
+[round(_basePrice * _vehiclepermanentbuy)] call life_fnc_numberText,
+_maxspeed,
+_horsepower,
+_passengerseats,
 if(_trunkSpace == -1) then {"None"} else {_trunkSpace},
-_vehicleInfo select 12,
-_vehicleInfo select 9
+_fuel,
+_armor
 ];
 
 _ctrl = CONTROL(2300,2304);
